requirements.txt
openai>=0.27.0
gradio>=3.40.0
python-dotenv>=1.0.0
python-docx>=0.8.11
